+++ example/unified_multi_step_form_example.dart
@@ -1,6 +1,0 @@
-import 'package:unified_multi_step_form/unified_multi_step_form.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}