+++ lib/src/unified_multi_step_form_base.dart
@@ -1,6 +1,5 @@
-// TODO: Put public facing types in this file.
-
-/// Checks if you are awesome. Spoiler: you are.
-class Awesome {
-  bool get isAwesome => true;
-}
+export 'controllers/unified_multi_step_form_controller.dart';
+export 'models/multi_step_form_page.dart';
+export 'theme/unified_multi_step_form_theme.dart';
+export 'utils/multi_step_form_validators.dart';
+export 'widgets/unified_multi_step_form.dart';