+++ lib/unified_multi_step_form.dart
@@ -1,8 +1,3 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library;
 
-export 'src/unified_multi_step_form_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/multi_form.dart';